requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pyyaml>=6.0

# PyTorch with MPS support (for Mac)
torch>=2.0.0

# Utilities
tqdm>=4.65.0
pydantic>=2.0.0
loguru>=0.7.0
joblib>=1.3.0
scipy>=1.11.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=11.0.0
kafka-python>=2.0.2
